Added selected place repo (still in memory)

+++ feature-main/src/main/java/se/gustavkarlsson/skylight/android/feature/main/gui/drawer/DrawerViewModel.kt
@@ -4,55 +4,71 @@ import androidx.annotation.DrawableRes
 import androidx.lifecycle.ViewModel
 import com.ioki.textref.TextRef
 import com.jakewharton.rxrelay2.PublishRelay
-import de.halfbit.knot.Knot
 import io.reactivex.Observable
-import se.gustavkarlsson.skylight.android.feature.main.Change
-import se.gustavkarlsson.skylight.android.entities.Place
-import se.gustavkarlsson.skylight.android.feature.main.State
+import io.reactivex.Scheduler
+import io.reactivex.android.schedulers.AndroidSchedulers
+import io.reactivex.rxkotlin.Observables
 import se.gustavkarlsson.skylight.android.feature.main.R
 import se.gustavkarlsson.skylight.android.lib.navigation.NavItem
 import se.gustavkarlsson.skylight.android.lib.navigation.Navigator
+import se.gustavkarlsson.skylight.android.lib.places.Place
 import se.gustavkarlsson.skylight.android.lib.places.PlacesRepository
+import se.gustavkarlsson.skylight.android.lib.places.SelectedPlaceRepository
 
 internal class DrawerViewModel(
-	private val mainKnot: Knot<State, Change>,
 	private val navigator: Navigator,
-	private val placesRepository: PlacesRepository
+	private val placesRepository: PlacesRepository,
+	private val selectedPlaceRepo: SelectedPlaceRepository,
+	observeScheduler: Scheduler = AndroidSchedulers.mainThread()
 ) : ViewModel() {
 	val places: Observable<List<DrawerItem>> =
-		mainKnot.state
-			.map(::createPlaceItems)
-			.distinctUntilChanged()
+		Observables.combineLatest(
+			placesRepository.all.toObservable(),
+			selectedPlaceRepo.selected,
+			::createPlaceItems
+		).observeOn(observeScheduler)
 
 	private val closeDrawerRelay = PublishRelay.create<Unit>()
 	val closeDrawer: Observable<Unit> = closeDrawerRelay
 
 	private val openRemoveLocationDialogRelay = PublishRelay.create<RemoveLocationDialogData>()
-	val openRemoveLocationDialog: Observable<RemoveLocationDialogData> = openRemoveLocationDialogRelay
+	val openRemoveLocationDialog: Observable<RemoveLocationDialogData> =
+		openRemoveLocationDialogRelay
 
-	private fun createPlaceItems(state: State): List<DrawerItem> =
-		state.places.map { place ->
-			val isActive = place == state.selectedPlace
-			val icon = when (place) {
-				is Place.Current -> R.drawable.ic_location_on_white_24dp
-				is Place.Custom -> R.drawable.ic_map_white_24dp
-			}
-			val onClick = {
-				mainKnot.change.accept(Change.SelectPlace(place))
-				closeDrawerRelay.accept(Unit)
-			}
-			val onLongClick: () -> Unit = if (place is Place.Custom) {
-				{
-					val dialogData = RemoveLocationDialogData(TextRef(R.string.remove_thing, place.name)) {
+	private fun createPlaceItems(places: List<Place>, selected: Place) =
+		places.map { place ->
+			DrawerItem(
+				isActive = place == selected,
+				icon = createIcon(place),
+				name = place.name,
+				onClick = createOnClick(place),
+				onLongClick = createOnLongClick(place)
+			)
+		} + createAddPlaceItem()
+
+	private fun createIcon(place: Place) =
+		when (place) {
+			is Place.Current -> R.drawable.ic_location_on_white_24dp
+			is Place.Custom -> R.drawable.ic_map_white_24dp
+		}
+
+	private fun createOnClick(place: Place): () -> Unit = {
+		selectedPlaceRepo.set(place)
+		closeDrawerRelay.accept(Unit)
+	}
+
+	private fun createOnLongClick(place: Place): () -> Unit =
+		if (place is Place.Custom) {
+			{
+				val dialogData =
+					RemoveLocationDialogData(TextRef(R.string.remove_thing, place.name)) {
 						placesRepository.remove(place.id)
 					}
-					openRemoveLocationDialogRelay.accept(dialogData)
-				}
-			} else {
-				{}
+				openRemoveLocationDialogRelay.accept(dialogData)
 			}
-			DrawerItem(isActive, icon, place.name, onClick, onLongClick)
-		} + createAddPlaceItem()
+		} else {
+			{}
+		}
 
 	private fun createAddPlaceItem(): DrawerItem {
 		val onClick = {
@@ -60,7 +76,13 @@ internal class DrawerViewModel(
 			navigator.push(NavItem("addplace"))
 		}
 		val onLongClick = {}
-		return DrawerItem(false, R.drawable.ic_add_white_24dp, TextRef(R.string.add_place), onClick, onLongClick)
+		return DrawerItem(
+			false,
+			R.drawable.ic_add_white_24dp,
+			TextRef(R.string.add_place),
+			onClick,
+			onLongClick
+		)
 	}
 }
